l3d: Use l3d lib for loading l3ds in g3d

+++ components/l3d/src/l3d_file.cpp
@@ -146,7 +146,7 @@ struct membuf: std::streambuf
 	{
 		if (way == std::ios_base::cur)
 		{
-			gbump(off);
+			gbump(static_cast<int>(off));
 		}
 		else if (way == std::ios_base::end)
 		{
